Move PySerializerInterface to module_py.cc
